# docs/swagger.yaml
basePath: /
definitions:
...
  models.BidangStudi:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.Book:
    properties:
      bidang_studi:
        $ref: '#/definitions/models.BidangStudi'
      bidang_studi_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      jenis_buku:
        $ref: '#/definitions/models.JenisBuku'
      jenis_buku_id:
        type: string
      jenjang_studi:
        $ref: '#/definitions/models.JenjangStudi'
      jenjang_studi_id:
        type: string
      kelas:
        $ref: '#/definitions/models.Kelas'
      kelas_id:
        type: string
      name:
        type: string
      price:
        type: number
      publisher:
        $ref: '#/definitions/models.Publisher'
      publisher_id:
        type: string
      updated_at:
        type: string
      year:
        type: string
    type: object
  models.City:
    properties:
      code:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.JenisBuku:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.JenjangStudi:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      period:
        type: string
      updated_at:
        type: string
    type: object
  models.Kelas:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Publisher:
    properties:
      address:
        type: string
      area:
        type: string
      city:
        $ref: '#/definitions/models.City'
      city_id:
        type: string
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone1:
        type: string
      phone2:
        type: string
      updated_at:
        type: string
      website:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@pustaka.com
    name: API Support
  description: REST API for Pustaka Digital Library Management System
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Pustaka Digital Backend API
  version: "1.0"
paths:
...
  /api/books:
    get:
      consumes:
      - application/json
      description: Retrieve all books with their related entities
      produces:
      - application/json
      responses:
        "200":
          description: List of all books
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all books
      tags:
      - Books
    post:
      consumes:
      - application/json
      description: Create a new book entry
      parameters:
      - description: Book details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Book'
      produces:
      - application/json
      responses:
        "201":
          description: Created book
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new book
      tags:
      - Books
  /api/books/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a book by ID
      parameters:
      - description: Book ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Book deleted successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Book not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a book
      tags:
      - Books
    get:
      consumes:
      - application/json
      description: Retrieve a single book by its ID with all related entities
      parameters:
      - description: Book ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Book details
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Book not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a book by ID
      tags:
      - Books
    put:
      consumes:
      - application/json
      description: Update an existing book by ID
      parameters:
      - description: Book ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated book details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Book'
      produces:
      - application/json
      responses:
        "200":
          description: Updated book
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Book not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a book
      tags:
      - Books
  /api/cities:
    get:
      consumes:
      - application/json
      description: Retrieve all cities ordered by creation date
      produces:
      - application/json
      responses:
        "200":
          description: List of all cities
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all cities
      tags:
      - Cities
    post:
      consumes:
      - application/json
      description: Create a new city entry
      parameters:
      - description: City details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.City'
      produces:
      - application/json
      responses:
        "201":
          description: Created city
          schema:
            $ref: '#/definitions/models.City'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new city
      tags:
      - Cities
  /api/cities/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a city by ID
      parameters:
      - description: City ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: City deleted successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: City not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a city
      tags:
      - Cities
    get:
      consumes:
      - application/json
      description: Retrieve a single city by its ID
      parameters:
      - description: City ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: City details
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: City not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a city by ID
      tags:
      - Cities
    put:
      consumes:
      - application/json
      description: Update an existing city by ID
      parameters:
      - description: City ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated city details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.City'
      produces:
      - application/json
      responses:
        "200":
          description: Updated city
          schema:
            $ref: '#/definitions/models.City'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: City not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a city
      tags:
      - Cities
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"